Refactor ReceptionAndAssignment in investigation cases
