+++ apps/judicial-system/web/messages/index.ts
@@ -21,7 +21,6 @@ export * from './InvestigationCases/Court/courtRecordForm'
 export * from './InvestigationCases/Court/ruling'
 export * from './InvestigationCases/Court/confirmation'
 export * from './InvestigationCases/Court/overview'
-export * from './InvestigationCases/Court/receptionAndAssignment'
 export * from './Indictments/Prosecutor/defendant'
 export * from './Indictments/Prosecutor/processing'
 export * from './Core'
