+++ apps/judicial-system/web/src/routes/Court/InvestigationCase/ReceptionAndAssignment/ReceptionAndAssignment.strings.ts
@@ -1,0 +1,10 @@
+import { defineMessages } from 'react-intl'
+
+export const icReceptionAndAssignment = defineMessages({
+  title: {
+    id: 'judicial.system.investigation_cases:reception_and_assignment.title',
+    defaultMessage: 'Móttaka og úthlutun',
+    description:
+      'Notaður sem titill á Móttaka og úthlutun skrefi í gæsluvarðhalds- og farbannsmálum.',
+  },
+})
